Add project settings and lower Java requirement to version 6

+++ org.eclipse.dawnsci.analysis.dataset.test/src/org/eclipse/dawnsci/analysis/dataset/LazyDynamicLoaderTest.java
@@ -93,7 +93,7 @@ public class LazyDynamicLoaderTest {
 			slice.setSlice(0, shape[0]-1, shape[0], 1);
 			slice.setSlice(1, shape[1]-1, shape[1], 1);
 			IDataset s = dataset.getSlice(slice);
-			Assert.assertEquals(max,(long)s.max());
+			Assert.assertEquals((long) max, s.max().longValue());
 			max+=400;
 			
 			AxesMetadata axm = s.getFirstMetadata(AxesMetadata.class);
